+++ previous-versions/accounting/src/sdk/models/operations/deletejournalentry.ts
@@ -7,20 +7,29 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class DeleteJournalEntryRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
     /**
-     * Unique identifier for a journal entry
+     * Unique identifier for a journal entry.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=journalEntryId" })
     journalEntryId: string;
 }
 
 export class DeleteJournalEntryResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -36,9 +45,15 @@ export class DeleteJournalEntryResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     pushOperationSummary?: shared.PushOperationSummary;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }