+++ previous-versions/accounting/src/sdk/models/operations/getpayment.ts
@@ -7,14 +7,23 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class GetPaymentRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a payment.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=paymentId" })
     paymentId: string;
 }
 
 export class GetPaymentResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -30,9 +39,15 @@ export class GetPaymentResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     payment?: shared.Payment;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }