+++ previous-versions/accounting/src/sdk/models/operations/listsalesorders.ts
@@ -7,6 +7,9 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class ListSalesOrdersRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
@@ -36,6 +39,9 @@ export class ListSalesOrdersRequest extends SpeakeasyBase {
 }
 
 export class ListSalesOrdersResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -51,9 +57,15 @@ export class ListSalesOrdersResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     salesOrders?: shared.SalesOrders;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }