+++ previous-versions/accounting/src/sdk/models/operations/updatepurchaseorder.ts
@@ -10,9 +10,15 @@ export class UpdatePurchaseOrderRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "request, media_type=application/json" })
     purchaseOrder?: shared.PurchaseOrder;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
@@ -22,14 +28,23 @@ export class UpdatePurchaseOrderRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=forceUpdate" })
     forceUpdate?: boolean;
 
+    /**
+     * Unique identifier for a purchase order.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=purchaseOrderId" })
     purchaseOrderId: string;
 
+    /**
+     * Time limit for the push operation to complete before it is timed out.
+     */
     @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=timeoutInMinutes" })
     timeoutInMinutes?: number;
 }
 
 export class UpdatePurchaseOrderResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -39,9 +54,15 @@ export class UpdatePurchaseOrderResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;